// src/app/ActivosComponent/models/direccion-activo.model.ts
export interface UbicacionActivoModel {
    nombreAula: string;
    nombreBloque: string;
    calleDireccion: string;
    detalleDireccion: string;
    zonaDireccion: string;
    nombreMunicipio: string;
    nombreProvincia: string;
    nombreDepartamento: string;
    nombrePais: string;
    nombreSucursal: string;
  }
  